+++ src/test/scala/JumpGroupInstructionsTest.scala
@@ -2,13 +2,15 @@ package net.mcribbs.s8008
 
 import org.scalatest.*
 import org.scalatest.wordspec.*
+import spire.implicits.ub
+import spire.math.UByte
 
 class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JMP" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x44)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x44))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       val state = cpu.step
       assert(state.PC == 0x0204)
@@ -17,8 +19,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JFC" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if C is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x40)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x40))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(carry = false))
       val state = cpu.step
@@ -26,8 +28,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if C is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x40)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x40))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(carry = true))
       val state = cpu.step
@@ -37,8 +39,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JFZ" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if Z is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x48)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x48))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(zero = false))
       val state = cpu.step
@@ -46,8 +48,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if Z is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x48)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x48))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(zero = true))
       val state = cpu.step
@@ -57,8 +59,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JFS" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if S is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x50)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x50))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(sign = false))
       val state = cpu.step
@@ -66,8 +68,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if S is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x50)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x50))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(sign = true))
       val state = cpu.step
@@ -77,8 +79,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JFP" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if P is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x58)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x58))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(parity = false))
       val state = cpu.step
@@ -86,8 +88,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if P is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x58)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x58))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(parity = true))
       val state = cpu.step
@@ -97,8 +99,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JC" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if C is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x60)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x60))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(carry = true))
       val state = cpu.step
@@ -106,8 +108,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if C is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x60)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x60))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(carry = false))
       val state = cpu.step
@@ -117,8 +119,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JZ" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if Z is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x68)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x68))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(zero = true))
       val state = cpu.step
@@ -126,8 +128,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if Z is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x68)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x68))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(zero = false))
       val state = cpu.step
@@ -137,8 +139,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JS" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if S is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x70)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x70))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(sign = true))
       val state = cpu.step
@@ -146,8 +148,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if S is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x70)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x70))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(sign = false))
       val state = cpu.step
@@ -157,8 +159,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
 
   "JP" should {
     "set the PC (unconditionally jump) to memory address B3...B3B2...B2 if P is true" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x78)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x78))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(parity = true))
       val state = cpu.step
@@ -166,8 +168,8 @@ class JumpGroupInstructionsTest extends AnyWordSpec {
     }
 
     "but not if P is false" in {
-      val ram = Memory(new Array[Byte](Memory.MAX_MEMORY)).writeByte(0x0000, 0x78)
-        .writeByte(0x0001, 0x04).writeByte(0x0002, 0x02)
+      val ram = Memory(new Array[UByte](Memory.MAX_MEMORY)).writeByte(0x0000, UByte(0x78))
+        .writeByte(0x0001, UByte(0x04)).writeByte(0x0002, UByte(0x02))
       val cpu = new CPU(ram)
       cpu.state = cpu.state.copy(flags = cpu.state.flags.copy(parity = false))
       val state = cpu.step
